Move PackageCodeTypeSubstate
