+++ radix-engine/src/kernel/interpreters/scrypto_interpreter.rs
@@ -16,7 +16,7 @@ use crate::system::node_modules::metadata::MetadataNativePackage;
 use crate::system::node_modules::royalty::RoyaltyNativePackage;
 use crate::system::node_modules::type_info::TypeInfoBlueprint;
 use crate::system::package::Package;
-use crate::system::type_info::PackageCodeTypeSubstate;
+use crate::system::package::PackageCodeTypeSubstate;
 use crate::types::*;
 use crate::wasm::{WasmEngine, WasmInstance, WasmInstrumenter, WasmMeteringConfig, WasmRuntime};
 use radix_engine_interface::api::node_modules::auth::{
