+++ radix-engine/src/system/mod.rs
@@ -5,4 +5,3 @@ pub mod node_modules;
 pub mod node_properties;
 pub mod node_substates;
 pub mod package;
-pub mod type_info;
