+++ radix-engine/src/system/node.rs
@@ -4,7 +4,7 @@ use crate::system::node_modules::access_rules::*;
 use crate::system::node_modules::metadata::MetadataSubstate;
 use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::node_substates::*;
-use crate::system::type_info::PackageCodeTypeSubstate;
+use crate::system::package::PackageCodeTypeSubstate;
 use crate::types::*;
 use radix_engine_interface::api::component::*;
 use radix_engine_interface::api::package::*;
