+++ radix-engine/src/system/package/mod.rs
@@ -1,5 +1,7 @@
 mod package;
 mod package_extractor;
+mod substates;
 
 pub use package::*;
 pub use package_extractor::*;
+pub use substates::*;
