+++ radix-engine/src/system/package/package.rs
@@ -8,7 +8,7 @@ use crate::system::node_modules::access_rules::{
 };
 use crate::system::node_modules::metadata::MetadataSubstate;
 use crate::system::node_modules::type_info::TypeInfoSubstate;
-use crate::system::type_info::PackageCodeTypeSubstate;
+use crate::system::package::PackageCodeTypeSubstate;
 use crate::types::*;
 use crate::wasm::{PrepareError, WasmValidator};
 use core::fmt::Debug;
