--- radix-engine/src/system/type_info/mod.rs
@@ -1,3 +1,0 @@
-mod substates;
-
-pub use substates::*;
